refactor log_rate_analysis route
