+++ x-pack/plugins/aiops/server/routes/categorization_field_validation/route_handler_factory.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import {
+import type {
   KibanaRequest,
   RequestHandlerContext,
   RequestHandler,
